Refactor: 관심사에 해당하는 container로 state 관리
App에서 모든 state관리를 담당하는 것은 관심사의 분리에 맞지 않기 때문에 각 관심사에 해당하는 container로 비즈니스 로직을 옮겼습니다.
